requirements.txt
asgiref==3.11.0
django==5.2.8
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.16.1
dotenv==0.9.9
et-xmlfile==2.0.0
json-log-formatter==1.1.1
openpyxl==3.1.5
pip==25.3
gunicorn==23.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.4
types-pyyaml==6.0.12.20250915
typing-extensions==4.15.0
uv==0.9.14
xlrd==2.0.2
